* commit '9c547a4ed96a31a5839ffbf14cc607ae8162e695':
  Cambio en categorias, proceso de paginacion vinculacion y desvinculacion

+++ src/Domain/Models/Categoria.cs
@@ -35,7 +35,7 @@ namespace Domain.Models
         public TipoCategoria TipoCategoria { get; set; }
         //
 
-        [Column("ORDEN", TypeName = "tinyint")]
+        [Column("ORDEN", TypeName = "int")]
         public int orden { get; set; }
 
         [Column("ICONO", TypeName = "varchar(200)")]
